convert some component to stateless functions
