+++ package.json
@@ -8,7 +8,7 @@
     "babel": "babel",
     "build": "webpack",
     "start": "npm run build && node index.js",
-    "coveralls": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls"
+    "coveralls": "cat ./coverage/lcov.info | coveralls"
   },
   "author": "Ajaps Franklin",
   "license": "ISC",
