+++ src/component/Footer.jsx
@@ -1,15 +1,7 @@
 import React from 'react';
 
-/**
- * Represents the Footer section.
- */
-export default class Footer extends React.Component {
-  /**
- * @returns {component} A component with Footer details.
- */
-  render() {
-    return (
+const Footer = () => (
     <footer className="footer">Headline News App (c) 2017 </footer>
-    );
-  }
-}
+  );
+
+export default Footer;
