+++ src/component/SourcesComponent.jsx
@@ -26,10 +26,11 @@ export default class SourcesComponent extends React.Component {
     // sets class name based on user source selection
     const btnClass = classNames({
       sourceBorder: this.props.sourceHigh[0] === this.props.id,
+      'sr-only': 'sr-only'
     });
 
     return (
-      <li className={btnClass}>
+      <li className={btnClass} >
         <a onClick={this.setSources.bind(this, this.props.id, this.props.sortBysAvailable)}>
           {this.props.name}
         </a>
