+++ src/stores/Article.js
@@ -18,10 +18,10 @@ class AllArticle extends EventEmitter {
       sortBy: 'top',
       source: 'all',
       URL_ARTICLES: 'https://newsapi.org/v1/articles',
-      API_KEY: '213327409d384371851777e7c7f78dfe',
+      API_KEY: process.env.NEWS_API_KEY,
     };
     this.status = 'ok';
-    this.sortAvailable = ['unavailable'];
+    this.sortAvailable = ['top'];
     this.highlightSource = ['all'];
     this.article = [];
   }
