+++ src/style/mainStyle.scss
@@ -57,6 +57,7 @@ a {
 .sourceBorder{
   border: solid black 1px;
   background: lightgoldenrodyellow;
+  border-radius: 1vw;
 }
 .caroselTitle{
   color: gold;
